+++ .scripts/build.sh
@@ -23,10 +23,10 @@ robocopy "$curDir/RCLootCouncil/Libs/" "$curDir/.tmp/RCLootCouncil_Classic/RCLoo
 # Build for Classic Era
 # -d: Skip Upload
 # -z: Skip zip
-"$release_script" -o -g classic -r "$curDir/.tmp" -m ".pkgmeta-build"
+"$release_script" -od -g classic -r "$curDir/.tmp" -m ".pkgmeta-build"
 
 # Build for BBC
-"$release_script" -o -g bcc -r "$curDir/.tmp" -m ".pkgmeta-build"
+"$release_script" -do -g bcc -r "$curDir/.tmp" -m ".pkgmeta-build"
 
 
 # # Move the zip