+++ Changelog.md
@@ -1,9 +1,21 @@
 ## Changes
 
-### Classic
+### Always auto award list
 
-Updated for patch 1.14.1.
+Added an option to choose items that should always be auto awarded. Located in the 'Awards' tab of the Master Looter options. Items can be added by itemID, name or item link.
+
+### TradeUI
+
+Implemented hotfix for handling duplicate items.
+Not as robust as I would like, but it should get the job done for now.
 
 ### Burning Crusade Classic
 
-Updated for patch 2.5.2.
+Updated .toc for patch 2.5.3.
+
+Added `Heart of Darkness` and `Mark of the Illidari` to the rep items loot list.
+
+
+### Bugfixes
+
+* *Fixed issue with `/rc remove` command.*